pyproject.toml
[tool.poetry]
name = "binance-coin-trading"
version = "0.1.0"
description = ""
authors = ["Bill1907 <a31829660@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "^1.0.19"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
pandas-ta = "^0.3.14b0"
matplotlib = "^3.9.0"
scikit-learn = "^1.5.0"
tensorflow = "^2.16.1"
schedule = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
